Handle ongoing work experience dates
