+++ src/components/ThemeContext.ts
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+export interface ThemeContextType {
+  theme: string;
+  setTheme: (theme: string) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
