+++ src/components/ThemeProvider.tsx
@@ -1,26 +1,16 @@
-// src/components/ThemeProvider.tsx
-'use client'; // <-- บอก Next.js ว่านี่คือ Client Component
+'use client';
 
-import { createContext, useContext, useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import type { ReactNode } from 'react';
+import { ThemeContext } from './ThemeContext';
 
-// สร้าง Context สำหรับ Theme
-interface ThemeContextType {
-  theme: string;
-  setTheme: (theme: string) => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-// สร้าง Provider Component
 export function ThemeProvider({ children }: { children: ReactNode }) {
-  const [theme, setThemeState] = useState('light'); // default theme
+  const [theme, setThemeState] = useState('light');
 
-  // Effect นี้จะทำงานเฉพาะบน Client เท่านั้น
   useEffect(() => {
     const savedTheme = localStorage.getItem('theme');
     const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-    
+
     if (savedTheme === 'dark' || (!savedTheme && prefersDark)) {
       setTheme('dark');
     } else {
@@ -46,11 +36,3 @@ export function ThemeProvider({ children }: { children: ReactNode }) {
   );
 }
 
-// สร้าง Custom Hook เพื่อให้เรียกใช้ง่าย
-export function useTheme() {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return context;
-}
